agency, auth, booking, payment, profile, rating, saved tour, tour,  controller finished and service repo modified

+++ src/main/java/quvoncuz/dto/tour/CreateTourRequestDTO.java
@@ -2,7 +2,6 @@ package quvoncuz.dto.tour;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -10,7 +9,7 @@ public class CreateTourRequestDTO {
     private String title;
     private String description;
     private String destination;
-    private BigDecimal price;
+    private Long price;
     private Integer durationDays;
     private Integer maxSeats;
     private LocalDate startDate;
